requirements.txt
fastapi==0.110.0
discord.py==2.3.2
pydantic==2.6.4
pydantic-settings==2.2.1
motor==3.3.2
pyjwt==2.8.0
uvicorn==0.29.0
httpx==0.27.0 